+++ docs/source/changes.rst
@@ -9,7 +9,8 @@ Release 0.043 - (under development)
 * Add more wrappers for CUBLAS 5 functions (enh. by Teodor Moldovan)
 * Add support for CULA Dense Free R17 (enh. by Alex Rubinsteyn)
 * Memoize elementwise kernel used by ifft scaling (#37).
-* Speed up maxabs function using reduction and kernel memoization.
+* Speed up misc.maxabs using reduction and kernel memoization.
+* Speed up misc.cumsum using scan and kernel memoization.
 
 Release 0.042 - (March 10, 2013)
 --------------------------------